requirements.txt
cooler==0.10.4
h5py==3.14.0
lightning==2.4.0
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
pyBigWig==0.3.24
pytorch_lightning==2.5.5
PyYAML==6.0.3
Requests==2.32.5
scikit_learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
torch==2.5.1
tqdm==4.67.1